Refactor Discover: extract sub-components and deduplicate shared logic
Extract SearchBar, ItemGrid, Tabs, EmptyStates into separate files.
Create discoverUtils with shared chipClass, restoreModalFromUrl, and
loadManifests helpers. Remove unused debounceRef, memoize showBadges,
improve error handling in handleStreamClick with error message display.
